Move wishlist content to the separate component
